refactor: RestaurantVideoHistoryFetcherException -> VideoHistoryFetcherException 예외명 변경 (#110)
